Removed "include depend" directive in makefile.x.
This was causing the build to fail.  Prerequisites are generated
automatically by GCC and makefile.x is rewritten by make.  So this
directive isn't needed anymore.

"With old make programs, it was traditional practice to use this
compiler feature to generate prerequisites on demand with a command
like `make depend'. That command would create a file depend
containing all the automatically-generated prerequisites; then the
makefile could use include to read them in (see Include).

"In GNU make, the feature of remaking makefiles makes this practice
obsolete—you need never tell make explicitly to regenerate the
prerequisites, because it always regenerates any makefile that is
out of date. See Remaking Makefiles."

- gnu.org/software/automake/manual/make/Automatic-Prerequisites.html