Rewrote config.c to use UNIX filenames
config.c was using dir.h from DJGPP and this caused the compile to
fail.  This include was removed and config.c had parts of the
savestatus() function rewritten to accomodate UNIX filenames.  This
currently breaks the DOS support.

The fnsplit() and fnmerge() functions were used to parse the drive
letter, path, basename and extension from the l->filename string. The
reason for doing so was to replace basenames and extensions.  Instead
of doing this in the function, linux_change_ext() and
linux_change_basename() were written to do this.  These should probably
be inplace edits and will be refactored later.

Tests were added using libtap in test_linux_config.c.  These test
the above functions and their helper functions.