convtabl.ini's path no longer overwritten
readpalettes() would bitwise-AND the address of some recently
malloc()'d memory with 0xffffff00 and use that result to store palette
data?  The number has no comment associated with it, and I don't know
what it does.  Commenting it out allows the conversion table to load
and doesn't throw up a config menu for Devil.