Rebuilt history
The repository now has the old source code on a different branch.
When all of the source code has been found, a new branch will be
created from the tip of the old source code branch and the commits
will be appended in order.

The script written to rebuild the history had trouble with dot-files.
Added the .gitignore back to the source.