Load all caps Descent 2 files
When loading the Descent 2 data, Devil was changed to load files with
all lowercase filenames.  The original Descent 2 data has filenames
in all capital letters, so the source code was changed again to load
these files.

Also, the default palette file in d2reg_11.ini had to be changed to
all caps as well, otherwise Devil would not load.